Bind Spacebar to Toggle, Improve Colors, Fix Mac OS Sierra Bug
The toggle is now bound to the spacebar and the segmentation lines
toggle to a pale gray. I also spread the colors apart a bit more so
hopefully they are easier to differentiate. There is a bug with Mac OS
Sierra that prevents imports from being in folders so for now
everything goes in root